+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/methods/engine/EngineGetPayloadV3.java
@@ -73,7 +73,7 @@ public class EngineGetPayloadV3 extends AbstractEngineGetPayload {
         CANCUN,
         cancunMilestone,
         PRAGUE,
-        protocolSchedule.flatMap(s -> s.milestoneFor(PRAGUE)),
+        maybeProtocolSchedule.flatMap(s -> s.milestoneFor(PRAGUE)),
         blockTimestamp);
   }
 }