+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/methods/engine/EngineNewPayloadV3.java
@@ -94,7 +94,7 @@ public class EngineNewPayloadV3 extends AbstractEngineNewPayload {
         CANCUN,
         cancunMilestone,
         PRAGUE,
-        protocolSchedule.flatMap(s -> s.milestoneFor(PRAGUE)),
+        maybeProtocolSchedule.flatMap(s -> s.milestoneFor(PRAGUE)),
         blockTimestamp);
   }
 }