+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/methods/engine/EngineNewPayloadV4.java
@@ -96,7 +96,7 @@ public class EngineNewPayloadV4 extends AbstractEngineNewPayload {
         PRAGUE,
         pragueMilestone,
         AMSTERDAM,
-        protocolSchedule.flatMap(s -> s.milestoneFor(AMSTERDAM)),
+        maybeProtocolSchedule.flatMap(s -> s.milestoneFor(AMSTERDAM)),
         blockTimestamp);
   }
 }