+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/response/RpcErrorType.java
@@ -61,6 +61,8 @@ public enum RpcErrorType implements RpcMethodError {
   INVALID_EXCESS_BLOB_GAS_PARAMS(
       INVALID_PARAMS_ERROR_CODE, "Invalid excess blob gas params (missing or invalid)"),
   INVALID_EXECUTION_REQUESTS_PARAMS(INVALID_PARAMS_ERROR_CODE, "Invalid execution requests params"),
+  INVALID_BLOCK_ACCESS_LIST_PARAMS(
+      INVALID_PARAMS_ERROR_CODE, "Invalid block access list params (missing or invalid)"),
   INVALID_SLOT_NUMBER_PARAMS(
       INVALID_PARAMS_ERROR_CODE, "Invalid slot number params (missing or invalid)"),
   INVALID_EXTRA_DATA_PARAMS(INVALID_PARAMS_ERROR_CODE, "Invalid extra data params"),